commit 9

+++ project/src/app/features/product-list/product-list.component.ts
@@ -1,8 +1,12 @@
-import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, ComponentFactoryResolver, ElementRef, OnInit, ViewChild, ViewContainerRef } from '@angular/core';
 import { Product } from 'src/app/shared/model/product.model';
 import { ProductListService } from './services/product-list.service';
-import { ProductManageFormComponent } from 'src/app/module/product-manage-form/product-manage-form.component';
+import { Currency1Pipe } from '../../shared/pipes/currency.pipe'
 import { Router } from '@angular/router';
+import { LazyLoadingService } from 'src/app/shared/LazyLoading/lazy-loading.service';
+import { CartService } from './services/cart.service';
+import { ProductManageFormComponent } from 'src/app/module/product-manage-form/product-manage-form.component';
+
 
 @Component({
   selector: 'app-product-list',
@@ -16,10 +20,14 @@ export class ProductListComponent implements OnInit,AfterViewInit {
   errorMessage = '';
   imageWidth = 50;
   imageMargin = 2;
+  color='white';
 
   constructor(private productService: ProductListService,
-              private router: Router) { }
-  @ViewChild(ProductManageFormComponent) addview !:ProductManageFormComponent
+              private router: Router,
+              private lazyLoadService: LazyLoadingService,
+              private cartService: CartService) { }
+  @ViewChild('manageFormContainerRef', {read: ViewContainerRef})
+  manageFormContainerRef!: ViewContainerRef;
   ngAfterViewInit(): void {
 
   }
@@ -28,27 +36,50 @@ export class ProductListComponent implements OnInit,AfterViewInit {
       next: products => {
         this.products= products;
         this.filteredProducts=products;
-        console.log(products);
         }
     })
-    
-    
   }
+  performFilter(filterBy: string) {
+    filterBy = filterBy.toLocaleLowerCase();
+    this.filteredProducts= this.products.filter((product: Product) =>
+    product.productName.toLocaleLowerCase().indexOf(filterBy) !== -1)
+     
+  }
+    
+  
   openEditForm(id: string){
-
-    this.addview.LoadEditData(id);
+    this.lazyLoadService.lazyload(this.manageFormContainerRef);
+    const compRef = this.manageFormContainerRef.createComponent(ProductManageFormComponent);
+    compRef.instance.LoadEditData(id);
+     
+   
 
   }
   deleteProduct(id: string){
-    this.productService.deleteProduct(id)
+    if (confirm("Do you want to save changes?") == true) {
+      this.productService.deleteProduct(id)
     .subscribe({
       next: () => {
         this.router.navigate(['home']);
       }
     });
+      
+  } else {
+      alert("delete cancelled");
+  }
+    
+  }
+  addToCart(product: Product) {
+    this.cartService.addToCart(product);
+    window.alert('Your product has been added to the cart!');
   }
 
+
   }
+
+
+
+
